comment out unneeded endpoint configs
there was code for the rawhid interface, which we don't really need,
so I commented that out. It still gets recognized by our code just
fine, but is otherwise untested.

We should rewrite all this into a neater package for our use, once
we are satisfied with it. It is no longer a rawhid device, so this
filename is a little bogus too...