sketch in the blocking read call
what we do is take a read request from a client, put the request struct
on a list for the read callback function to grab. That callback will
search the list for matching device numbers, copy the data to the buffer
in the read request, mark it as completed and then wakeup the readers
wait_queue. THat will unblock the read, which will check some validity
and return... flawless, right?!