implemented adc_read() that calls (no-op) teensy_read()
right now i have a dummy t_dev in the read_request, and i'm not sure
about the semantics of teensy_read()'s return value.  but i think this
completes #3 in andrew's email (the adc dev i said i'd implement).

as part of this commit, i moved some stuff from teensy.h into
teensy.c.  it was giving me errors and warnings, when i tried to
include teensy.h in teensy_adc.c, and was all marked static, so i
think i was correct to move it. but i don't understand why it was in
teensy.h in the first place, so i'm a little worried i'm messing with
something i don't understand ...