abstracted initialization and exit of submodules into functions
this was suggested by a comment in teensy.c.  i wanted the submodules lists to go in the header submmodules.h, but that leads to multiple definition errors because it's included in other files that get linked against submodules.o. a separate header file seemed like overkill ... i sux at the C